chore: dedupe DOM helpers and remove dead ai/ module
- Extract esperar()/simularCliqueReal()/simularClique() into a shared
  shared/dom-utils.js, removing 6 identical copies of esperar() and
  3 identical/near-identical copies of the click-simulation helpers
  scattered across utils.js, page-data.js, command-center.js,
  notes-bridge.js, case-log-scraper.js and email-automation-service.js.
- Remove src/modules/ai/ (gemini-service.js + magic-wand.js): confirmed
  unused, not imported from app.js or any other module, so it never
  even shipped in the bundle.

Pure dedup/dead-code removal, no behavior change. Verified with
esbuild --bundle --minify (same command CI uses) that the app still
builds cleanly after the changes.
